Update Double_DDS_VFO_0v2.ino
Declared the variable "Turns" as an int rather than as a double

+++ Double_DDS_VFO_0v2.ino
@@ -170,7 +170,7 @@ byte RiTSignPos=23;
 byte VFOPos=26;
 byte BandPos=28;
 
-double Turns; 
+int Turns;      // was previously declared as a double
 int Menu = 0;
  
 // end of Variables
